JG-50 Refactor
