+++ src/routes/card.ts
@@ -3,16 +3,14 @@
  */
 import * as express from "express";
 import credentials from "../credentials";
-import {ProjectService} from  "../service/projectService";
-import {CustomFieldService} from "../service/customFieldService";
-import {IssueService} from "../service/issueService";
-import {EpicService} from "../service/epicService";
+import {ProjectService} from "../service/projectService";
+import {EpicAsParentService} from "../service/epicAsParentService";
 
 export default (addon) => {
     const router = express.Router();
     const httpClient = addon.httpClient(credentials);
     const projectService = new ProjectService();
-    const epicService = new EpicService();
+    const epicAsParentService = new EpicAsParentService();
 
     router.get('/project', (req, res) => {
         projectService.getProjectCards(httpClient).then((projectResponse) => {
@@ -22,7 +20,7 @@ export default (addon) => {
     });
 
     router.get('/project/:key', (req, res) => {
-        epicService.getEpicsForProjectCards(req.params.key, httpClient).then((epicResponse) =>{
+        epicAsParentService.getEpicCardsForProjectKey(req.params.key, httpClient).then((epicResponse) => {
             res.setHeader("Content-Type", "application/json");
             res.send(epicResponse);
         });
